+++ connect-for-commerce-scope-module/src/main/java/to/be/renamed/dap/category/EcomCategoryDataStream.java
@@ -74,14 +74,17 @@ public class EcomCategoryDataStream implements DataStream<EcomCategory> {
 
     private EcomSearchResult<EcomCategory> getItems(Map<String, String> filters, int page) {
         try {
-            return ServiceFactory.getBridgeService(scope.getBroker()).findCategories(filters.get(EcomDapUtilities.FILTER_PARENT_ID), scope.getLang(), page);
+            return ServiceFactory.getBridgeService(scope.getBroker())
+                .findCategories(filters.get(EcomDapUtilities.FILTER_QUERY), filters.get(EcomDapUtilities.FILTER_PARENT_ID), scope.getLang(), page);
         } catch (BridgeConnectionException e) {
-            Logging.logError(format(EcomDapUtilities.ERROR_LOG_MESSAGE, EcomDapUtilities.ERROR_BRIDGE_CONNECTION, e.getErrorCode()), e, this.getClass());
+            Logging.logError(format(EcomDapUtilities.ERROR_LOG_MESSAGE, EcomDapUtilities.ERROR_BRIDGE_CONNECTION, e.getErrorCode()), e,
+                             this.getClass());
             EcomDapUtilities.openDialog(e.getLocalizedMessage(), e.getErrorCode(), scope);
             return new EcomSearchResult<>(Collections.emptyList(), 0);
         } catch (BridgeException e) {
             // Workaround for a bug in the Salesforce bridge
-            Logging.logError(format(EcomDapUtilities.ERROR_LOG_MESSAGE, EcomDapUtilities.ERROR_BRIDGE_CONNECTION, e.getErrorCode()), e, this.getClass());
+            Logging.logError(format(EcomDapUtilities.ERROR_LOG_MESSAGE, EcomDapUtilities.ERROR_BRIDGE_CONNECTION, e.getErrorCode()), e,
+                             this.getClass());
             total--;
             return new EcomSearchResult<>(Collections.emptyList(), 0);
         }