+++ connect-for-commerce-scope-server/src/main/java/to/be/renamed/bridge/DefaultBridgeService.java
@@ -70,13 +70,14 @@ public class DefaultBridgeService implements BridgeService {
     /**
      * Returns all available categories from the shop backend in the specified language.
      *
+     * @param q        Filter by category name. All categories which names match the query will be returned.
      * @param parentId Filter by parentId. It returns all related categories and subcategories from the category tree.
      * @param lang     The language of the result.
      * @param page     The current page of the search results.
      * @return The available categories from the shop backend in the specified language.
      */
-    public EcomSearchResult<EcomCategory> findCategories(@Nullable String parentId, @Nullable String lang, final int page) {
-        return bridgeFsService.findCategories(projectId, parentId, lang, page);
+    public EcomSearchResult<EcomCategory> findCategories(@Nullable String q, @Nullable String parentId, @Nullable String lang, final int page) {
+        return bridgeFsService.findCategories(projectId, q, parentId, lang, page);
     }
 
     /**